Added classes that implement bit based masks
Added a couple of classes that implement binary masks using
the individual bits of of bytes as booleans. These are added
so that I can make a boolean mask of an entire data set without
using an absurd amount of memory.

The classes are defined in:
	image/bitMask.h
	image/volumeMask.h
and implemented in the corresponding .cpp files.