Added a Blobmapper class
Added a class that identifies contiguous blobs as volumes of
increasing intensity. This works very well for finding small
peaks, but does not yet work great for finding nuclei as these
contain many sub-peaks.

I still need to add reasonable widgets to control the results
and functions to output results, and so on.