Implemented the data piping around the ND_Classifier and BlobClassifier
Implemented hooks and some interfaces to permit the use of the
classifier. Added one new class, the SuperBlobWidget for this,
and added a few functions in BlobMapperWidgetManager and the
BlobClassifier classes.

I had some trouble with the ND_Classifier class, and since
this uses rather low level datastructures combined with STL
maps in rather a too complex a way I expect that there will be
some more bugs to iron out.