Command line option for blobMapping / background estimation.
Added command line switches and appropriate plumbing to allow
blobmapping to be specified on the command line. This allows
a shell script to map blobs from a specified set of files,
with a resulting export of blob files.

Am halfway through creating a background estimating class.
(image/background.h), this uses simple linear interpolation
from a number of cubes of data. This appears to be broken by
rounding errors of various kinds.