Background based blob finding.
The find blobs function now uses a background estimate to distinguish
local peaks (in 3D hence blobs). Blobs are initiated wherever the
signal is higher than the background; however, only blobs that have
a maximal value that is higher than the background + value given
are kept. Voxels belonging to local peaks that don't qualify are
instead masked out. This is slower, but, does a better job of
compensating for a variable background.

The input window has also been modified to allow the user to specify
how the background is calculated, with the size of the x, y and z
windows used, as well as the percentile that is taken as the background
level being settable (as a long string of comma seperated values).

However, the command line option for mapping blobs has not yet been
modified to set these parameters, but will just take the default
parameters encoded in the source.