ssize_t changed to size_t
Realised that ssize_t indicates a signed integer, which on 32 bit
machines limits the size of files that can be looked at. Changed
all instances of ssize_t to the more reasonable size_t. Still
somewhat concerned that size_t is only 32 bits long on 32 bit
Linux (at least on Ubuntu). This seems a bit limiting.