Added a simple background subtraction algorithm for the viewer.
Added a simple background subtraction algorithm that works on a per
frame basis to reduce the background in viewed images. I should replace
this with something that uses the previous 3D background subtraction.

I also need to add commands to allow the parameters to be set for the
background, as presently it defaults to things which work well for small
spots, but terribly for big things like nuclei.

I also found that there is some buge that causes the fileName variable
in the FileSet object to be overwritten for some specific files. Not
sure what is causing this, but have put in hack to rescue the creation
of the .info file in the meanwhile.