Implemented an ImageBuilder
Impemented a few classes to build very specific image based on selected
stacks. This allows for example the building of a projection of from
a specified range of slices with or without background subtraction.

Also implemented two more command line options, -r and -c for specifying
a file containing the colors to be used, and one for the viewing ranges
to use for setting the bias and scale (contrast & brightness).

Didn't do much with the cavity mapper function. Hopefully around the
next commit.