In the middle of implementing a manual overlap editor
Halfway through making a manual editor for setting the overlaps between
panels. As part of this process I have modified the GLImage class such
that it now accepts different textureWidths and textureHeights. I've
also added a function within it that paints a bigger image to the panels
so that this doesn't need to be done externallly.