Implemented manual overlap editor and new overlap merge algorithm
Implemented a manual overlap editor that allows corrections in the
automated overlap alignments. The experience of using this suggests
strangely that previous errors primarily arose from the use of the
DAPI channel for alignments (leading to about 200 nm offsets).

Having a perfect alignment is not enough however, as the background
at the edges is quite different, perhaps reflecting the microscope
setup. Hence this commit also has the beginnings of a new overlap
merger algorithm that averages the overlapped region. That is quite
tricky though, as we sometimes do want to average, whereas sometimes
it's a bad idea. The large number of redundant read options don't
really help here, and the following commits will reflect an effort
to remove some old redundant read functions. Then we might be able
to do some much more clever things.