Added multithreaded blob mapper
Added a multithreaded blob mapper that allows blobs to be mapped from
subsets of the original image stack. The ImageBuilder function has
a function that creates a number of these to allow for accelerated blob
mapping on large data sets without running out of memory.

The mt_blob_mapper will be modififed in the next set of changes to make
the process of creating and destroying image stacks and associated
information more streamlined (this is necessary in large data sets in
order not to run out of memory).