Added a CellTracer class, and various small changes.
Included a CellTracer class which tries to work out the outlines of
cells using NNMap-ped blobs (converted to Point* objects). The class
doesn't actually trace outlines, but rather makes a cell mask using a
combination of point linkages and flood filling. It kind of works, but
not great yet.

Modified ImageBuilder::f_project such that one can project only the
currently visible area of the ImageBuilder area. This hooks into a
simple function in GLImage that informs the caller of the current area.
Should perhaps use also for the view window as this would be quite easy
to implement.