Added Cell editing functions
Completed the ability to edit cell boundaries in the main ImageBuilder
image window. This required the addition of a number of new classes as
well as new functions in ImageBuilder. The functions appear to work, but
could do with some refinements to make the process faster (in particular
only the data from one cell at a time is shown). Also need to make an
output function, and to update the ImageBuilder destructor which will be
sorely out of date.