New Blob Merger class and functions saving / reading CellCollections
Added a new BlobMerger class (a friend of blob_mt_mapper) that is used
inside of Blob_mt_mapper to merge blobs of different wavelengths. This
uses a new heuristic to assess the merging of blobs.

Added functions to allow the saving and reading to/from file of
CellCollection objects. At the moment, the functions do very little
error checking, but seem to work OK.

Note that the BlobMerger class uses more strict functions to avoid
merging blobs that shouldn't be merged. However, this makes it necessary
to correct for chromatic shift; although I have previously created
functions that allow (from the deltaviewer window) to set X-Y-Z offsets
for the individual channels; this unfortunately breaks the merger
function with a difficult to trace segmentation fault. I obviously need
to do some serious digging around in the code to work out where. It
looks like fixing this should fix a number of problems.