Added Self-organising-deltoids classes
Added the sod/ directory. This contains a number of classes that can be
used to visualise n-dimensional relationship in lower dimensions
(currently only 2-dimensional, but that should be easy to change). These
classes are currently a bit disorganised and require as input distance
matrices and optional position matrices (rather than generate the
distance matrices itself). This is because I used to use these functions
for visualising relationships in very high-dimensional data (i.e.
microarray data) and hence used it a little bit differently.

These classes have not been included into the main dvreader application,
but there is a separate .pro file and a separate executable. However,
since this uses classes defined in the main dvreader project I will
include it here. I may also include direct access to these classes from
the main application, but am not sure if this is a good idea at the
moment.