Updates to sod/ classes and margin to rolloff in FrameStack
Added the ability to draw points as pie charts for the self-organising
deltoids. Added the ability to define a grid in n-dimensional space in
order to plot the n-dimensional space deformation that results from the
rearrangement of points. However, this is terribly slow.

Removed the use of the margin variable from FrameStack. This has been
set to 0, and the xyMargin is instead set to the rolloff variable. This
can be set with the command line switch -m.