Added classes to allow multithreaded lookup tables
Added classes to allow short values to be converted to rgb floats by
simple lookup as opposed to the currently used direct calculation
(which includes lots of casts). This should allow a faster drawing
of stacks and slices; at least when the speed is not completely limited
by disk access.

Also managed to changed the spacing of the colour boxes.