Implemented multithreaded lookup tables for readToRGB function
Used the previously developed lookup table classes to implement a lookup
table for RGB values as opposed to direct calculation. So far this has
only been implemented when we don't perform background subtraction or
anything else complicated.

Note that there are still likely to be some bugs present in this code.