Added multithreaded Vector Adjustment
Added the VectorAdjustThread class to make the vector adjustment
multithreaded. Also modified the functions as to avoid a lot of
constructors and destructors. This allows the decrease of the mapping
time down to a fifth of the time taken previously.
(Though this will depend on the processor(s) available). Hyperthreading
doesn't seem to help.