Added the option not to remember and plot component vectors
Added the option not to remember the component vectors that make up the
force vector moving the dpoint. This Drastically speeds up the mapping
and now allows me to do 4000 points (using 4 dimensions and 1000
iterations) in 250 seconds. Previously this was not possible to do at
all due to the memory requirements of remember so many vectors.

This is set with the coords=true/false option for the read_distances
function. I am however planning to change this to a comp=true/false.