Various SOD function additions. PerimeterSplitter bug fix
Added triangle grid function to distanceMapper (and functions to plot
the resulting triangles) to visualise distortions in N-D space.

Fixed a long standing bug in the perimeter splitter function.
(Where if two nuclei were connected by a single pixel width line, the
top nucleus would disappear). This bug was due to the behaviuor of flood
filling and diagonals.