Several bug fixes
Fixed bug in perimeterPlotter display to allow drawing of background
(not sure why it stopped working). Also fixed a bug in BlobMerger that
caused the specified channel offsets to be ignored.

And most importantly, fixed the code in main.cpp such that the stack
size is automatically increased (in a not particularly smart way) to
avoid running out of stack in some of the recursive functions.