2 Bugfixes in GLImage, and console message removals
Fixed a memory segmentation bug in an unused part of the GLImage
initialisation. Also made sure that the viewingState variable is set in
both GLImage constructors (ought to rewrite that so that initialisation
is done only in one piece of code, but later maybe).

Also removed or commented out a load of superflous cout statements.