Base openCL class implemented
Implemented a base open CL class that sets up a single kernel from a
source file and initialises that to run on the GPU. Not terribly
flexible. I've folded the maximum intensity projection up into this and
this works tolerably well. Time constraints for simple things like this
seem to be related to copying data back and forth in memory. A MIP of
1024x1024x20 takes about 21 ms of which about 18 are spent applying
arguments (i.e. copying data across memory).