Bleach counting mechanisms
Added functions and classes to estimate the bleaching in a set of
panels. This is is a halfway commit for these functions and does not yet
seem to give the correct numbers.

Also modified the behaviour of glImage such that zooming can be done
with left-right as well as up-down mouse movement (right button).