Perimeter Class cleanup
cleaned up a few of the functions in the perimeter class in order to
remove friends of Perimeter. The Perimeter class is a total mess, but is
unfortunately used in many places making it difficult to get rid of it.
In order to make it safer for use, I'm removing friend classifications
that were introduced because I was too lazy to write a load of accessor
functions. This is in order to be able to make certain guarantees about
the properties of the perimeters, things like completeness.